requirements.txt
# SHC - Sparse Selective Hyper-Connections
# Requirements for Phase 1 Core Components

# Core dependencies
torch>=2.0.0
numpy>=1.24.0
einops>=0.6.0

# Training utilities
tqdm>=4.65.0
tensorboard>=2.12.0

# Configuration
pyyaml>=6.0
hydra-core>=1.3.0

# Distributed training
deepspeed>=0.9.0  # Optional: for efficient multi-GPU

# Memory profiling
memory-profiler>=0.61.0
psutil>=5.9.0

# Development
pytest>=7.0.0
pytest-benchmark>=4.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0
